+++ vlib/rand/xoroshiro128pp/xoros128pp_test.v
@@ -159,7 +159,7 @@ fn test_xoroshiro128pp_u64_in_range() {
 
 fn test_xoroshiro128pp_int31() {
 	max_u31 := int(0x7FFFFFFF)
-	sign_mask := int(0x80000000)
+	sign_mask := int(u32(0x80000000))
 	for seed in seeds {
 		mut rng := &rand.PRNG(&XOROS128PPRNG{})
 		rng.seed(seed)
@@ -175,7 +175,7 @@ fn test_xoroshiro128pp_int31() {
 
 fn test_xoroshiro128pp_int63() {
 	max_u63 := i64(0x7FFFFFFFFFFFFFFF)
-	sign_mask := i64(0x8000000000000000)
+	sign_mask := i64(u64(0x8000000000000000))
 	for seed in seeds {
 		mut rng := &rand.PRNG(&XOROS128PPRNG{})
 		rng.seed(seed)
